refactor: error handling
